Properly dequeue cancelled events when multiple apps are rendered (#2540)
* Fix render + tests

* Add comment + changelog
